Oglasi autorizacija

+++ app/Http/Controllers/OglasController.php
@@ -9,9 +9,16 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
+use JWTAuth;
+use Illuminate\Http\Response as HttpResponse;
 
 class OglasController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('jwt.auth');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -40,35 +47,46 @@ class OglasController extends Controller
      */
     public function store(Request $request)
     {
-        $oglas = new Oglas;
-        $oglas->naziv = $request->input('naziv');
-        $oglas->tip_oglasa = $request->input('tip');
-        $oglas->cijena = $request->input('cijena');
-        $oglas->povrsina = $request->input('povrsina');
-        $oglas->stanje = $request->input('stanje');
-        $oglas->opis = $request->input('opis');
+        try{
 
-        $lokacija = new Lokacija;
-        $lokacija->drzava = $request->input('drzava');
-        $lokacija->kanton = $request->input('kanton');
-        $lokacija->grad = $request->input('grad');
-        $lokacija->opstina = $request->input('opstina');
-        $lokacija->adresa = $request->input('adresa');
-        $lokacija->save();
-        $oglas->lokacija_id = $lokacija->id;
+            $token = JWTAuth::getToken();
+            $user = JWTAuth::toUser($token);
+            
+            
+            $oglas = new Oglas;
+            $oglas->naziv = $request->input('naziv');
+            $oglas->tip_oglasa = $request->input('tip');
+            $oglas->cijena = $request->input('cijena');
+            $oglas->povrsina = $request->input('povrsina');
+            $oglas->stanje = $request->input('stanje');
+            $oglas->opis = $request->input('opis');
 
-        $oglas->autor_id = $request->input('autor_id');
-        $oglas->datum_objave = Carbon::now();
+            $lokacija = new Lokacija;
+            $lokacija->drzava = $request->input('drzava');
+            $lokacija->kanton = $request->input('kanton');
+            $lokacija->grad = $request->input('grad');
+            $lokacija->opstina = $request->input('opstina');
+            $lokacija->adresa = $request->input('adresa');
+            $lokacija->save();
+            $oglas->lokacija_id = $lokacija->id;
 
-        $oglas->grijanje = $request->input('grijanje');
-        $oglas->struja = $request->input('struja');
-        $oglas->voda = $request->input('voda');
-        $oglas->telefon = $request->input('telefon');
-        $oglas->kablovska = $request->input('kablovska');
-        $oglas->internet = $request->input('internet');
-        $oglas->garaza = $request->input('garaza');
+            $oglas->autor_id = $user->id;
+            $oglas->datum_objave = Carbon::now();
 
-        $oglas->save();
+            $oglas->grijanje = $request->input('grijanje');
+            $oglas->struja = $request->input('struja');
+            $oglas->voda = $request->input('voda');
+            $oglas->telefon = $request->input('telefon');
+            $oglas->kablovska = $request->input('kablovska');
+            $oglas->internet = $request->input('internet');
+            $oglas->garaza = $request->input('garaza');
+
+            $oglas->save();
+
+        } catch(Exception $e){
+            return response()->json(['error' => 'Error storing oglas'], HttpResponse::HTTP_CONFLICT);
+        }
+
     }
 
     /**
@@ -102,88 +120,102 @@ class OglasController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $oglas = Oglas::find($id);
+        try{
+            $token = JWTAuth::getToken();
+            $user = JWTAuth::toUser($token);
+            
+            $oglas = Oglas::find($id);
+            
+            if($user->id == $oglas->autor->id || $user->admin){
+                if ($request->has('naziv')) {
+                    $oglas->naziv = $request->input('naziv');
+                }
 
-        if ($request->has('naziv')) {
-            $oglas->naziv = $request->input('naziv');
+                if ($request->has('tip')) {
+                    $oglas->tip_oglasa = $request->input('tip');
+                }
+
+                if ($request->has('status')) {
+                    $oglas->status_oglasa = $request->input('status');
+                }
+
+                if ($request->has('cijena')) {
+                    $oglas->cijena = $request->input('cijena');
+                }
+
+                if ($request->has('povrsina')) {
+                    $oglas->povrsina = $request->input('povrsina');
+                }
+
+                if ($request->has('stanje')) {
+                    $oglas->stanje = $request->input('stanje');
+                }
+
+                if ($request->has('opis')) {
+                    $oglas->opis = $request->input('opis');
+                }
+
+                $lokacija = $this->dajLokaciju($id);
+
+                if ($request->has('drzava')) {
+                    $lokacija->drzava = $request->input('drzava');
+                }
+
+                if ($request->has('kanton')) {
+                    $lokacija->kanton = $request->input('kanton');
+                }
+
+                if ($request->has('grad')) {
+                    $lokacija->grad = $request->input('grad');
+                }
+
+                if ($request->has('opstina')) {
+                    $lokacija->opstina = $request->input('opstina');
+                }
+
+                if ($request->has('adresa')) {
+                    $lokacija->adresa = $request->input('adresa');
+                }
+
+                if ($request->has('grijanje')) {
+                    $oglas->grijanje = $request->input('grijanje');
+                }
+
+                if ($request->has('struja')) {
+                    $oglas->struja = $request->input('struja');
+                }
+
+                if ($request->has('voda')) {
+                    $oglas->voda = $request->input('voda');
+                }
+
+                if ($request->has('telefon')) {
+                    $oglas->telefon = $request->input('telefon');
+                }
+
+                if ($request->has('kablovska')) {
+                    $oglas->kablovska = $request->input('kablovska');
+                }
+
+                if ($request->has('internet')) {
+                    $oglas->internet = $request->input('internet');
+                }
+
+                if ($request->has('garaza')) {
+                    $oglas->garaza = $request->input('garaza');
+                }
+
+                $lokacija->save();
+                $oglas->save();
+                return response()->json(['success' => 'Oglas info updated'], HttpResponse::HTTP_OK);
+            }
+            else return response()->json(['error' => 'No authorization to update'], HttpResponse::HTTP_UNAUTHORIZED);
+
+            
+        } catch(Exception $e){
+            return response()->json(['error' => 'Error storing oglas'], HttpResponse::H);
         }
-
-        if ($request->has('tip')) {
-            $oglas->tip_oglasa = $request->input('tip');
-        }
-
-        if ($request->has('status')) {
-            $oglas->status_oglasa = $request->input('status');
-        }
-
-        if ($request->has('cijena')) {
-            $oglas->cijena = $request->input('cijena');
-        }
-
-        if ($request->has('povrsina')) {
-            $oglas->povrsina = $request->input('povrsina');
-        }
-
-        if ($request->has('stanje')) {
-            $oglas->stanje = $request->input('stanje');
-        }
-
-        if ($request->has('opis')) {
-            $oglas->opis = $request->input('opis');
-        }
-
-        $lokacija = $this->dajLokaciju($id);
-
-        if ($request->has('drzava')) {
-            $lokacija->drzava = $request->input('drzava');
-        }
-
-        if ($request->has('kanton')) {
-            $lokacija->kanton = $request->input('kanton');
-        }
-
-        if ($request->has('grad')) {
-            $lokacija->grad = $request->input('grad');
-        }
-
-        if ($request->has('opstina')) {
-            $lokacija->opstina = $request->input('opstina');
-        }
-
-        if ($request->has('adresa')) {
-            $lokacija->adresa = $request->input('adresa');
-        }
-
-        if ($request->has('grijanje')) {
-            $oglas->grijanje = $request->input('grijanje');
-        }
-
-        if ($request->has('struja')) {
-            $oglas->struja = $request->input('struja');
-        }
-
-        if ($request->has('voda')) {
-            $oglas->voda = $request->input('voda');
-        }
-
-        if ($request->has('telefon')) {
-            $oglas->telefon = $request->input('telefon');
-        }
-
-        if ($request->has('kablovska')) {
-            $oglas->kablovska = $request->input('kablovska');
-        }
-
-        if ($request->has('internet')) {
-            $oglas->internet = $request->input('internet');
-        }
-
-        if ($request->has('garaza')) {
-            $oglas->garaza = $request->input('garaza');
-        }
-
-        $lokacija->save();
-        $oglas->save();
+        
 
     }
 
@@ -195,11 +227,20 @@ class OglasController extends Controller
      */
     public function destroy($id)
     {
-        $oglas = Oglas::find($id);
-        $oglas->lokacija()->delete();
-        $oglas->delete();
-    }
+        $token = JWTAuth::getToken();
+        $user = JWTAuth::toUser($token);
 
+        $oglas = Oglas::find($id);
+
+        if($user->id == $oglas->autor->id || $user->admin){
+            $oglas->lokacija()->delete();
+            $oglas->delete();
+            return response()->json(['success' => 'Oglas deleted'], HttpResponse::HTTP_OK);
+        }
+        else return response()->json(['error' => 'No authorization to delete'], HttpResponse::HTTP_UNAUTHORIZED);
+        
+        
+    }
     
     public function poTipuOglasa($tip)
     {
@@ -223,24 +264,41 @@ class OglasController extends Controller
 
     public function dodajSliku(Request $request,  $id)
     {
+        $token = JWTAuth::getToken();
+        $user = JWTAuth::toUser($token);
         
-        $slika = new Slika;
-        $slika->oglas_id = $id;
-        $slika->slika = $request->input('slika');
-        $slika->save();
-
+        $korisnik = Oglas::find($id)->autor;
+        
+        if($user->id == $korisnik->id || $user->admin){
+            $slika = new Slika;
+            $slika->oglas_id = $id;
+            $slika->slika = $request->input('slika');
+            $slika->save();
+            return response()->json(['success' => 'Slika saved'], HttpResponse::HTTP_OK);
+        }
+        else return response()->json(['error' => 'No authorization to add Slika'], HttpResponse::HTTP_UNAUTHORIZED);
+        
     }
     
     public function obrisiSliku($id, $slika_id)
     {
-        $slika = Slika::find($slika_id);
-        $slika->delete();
+        $token = JWTAuth::getToken();
+        $user = JWTAuth::toUser($token);
+
+        $korisnik = Oglas::find($id)->autor;
+
+        if($user->id == $korisnik->id || $user->admin){
+            $slika = Slika::find($slika_id);
+            $slika->delete();
+            return response()->json(['success' => 'Slika deleted'], HttpResponse::HTTP_OK);
+        }
+        else return response()->json(['error' => 'No authorization to delete Slika'], HttpResponse::HTTP_UNAUTHORIZED);
+        
     }
 
     public function dajFavorite($id)
     {
         return Oglas::find($id)->favoriti;
     }
-
-
+    
 }
